Feature/named pipes (#10)
* The beginning of named pipes.

* Maturing named pipes

* Failing unit test

* Maturing Named Pipes

* Updating documentation.

* Maturing named pipe tests

* Fixing failing test.

* Unit tests matured. cleaning syntax next.

* Started a regression test harness. Matured test coverage.

Coverage is not at 100%, but it's a catch block that's not officially tested. Will likely fix.

* Partial regression named pipe working

Server will start and receive a message. Responding is not working.

* Adding server-to-client response

Still not fully working. Named pipes are still throwing an exception on the response side of things. Looking into this next.

* Refactoring named pipes

* Syntax cleanup

* Matured named pipes, bad loop intro'd server side

Need to get rid of the bad loop, obviously.

* Server wrapper causing odd loop behavior, I think.

* Found a couple of errors I made in my tired code.

* Single message is working. Perpetual is not.

* Improving logging.

* Still fighting the second message send issue.

* Figured it out!

There were a few errors in my code, chief of which was that the client was ALREADY CONNECTED after the first message send. Now it checks to see if the client is already connected and skips over the stream connection process if so.

* Adding a TODO.

* Adding disconnect

* Improving test coverage

* Maturing tests again.

* Only exceptions are uncovered, which is fine because I'm returning those messages to the consumer; this library is not throwing exceptions for named pipes.